backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
psycopg2-binary
pandas
numpy
xgboost
python-dotenv
